Refactor Authorization.
